network code
-- fix (peerout can read data from server)
-- need to test (read data from peerout)

+++ Freechat/Network/peerout.h
@@ -32,6 +32,6 @@ public slots:
 
 private:
     QTcpSocket *socket;
-    quint16 nextBlockSize;
+    qint64 nextBlockSize;
 };
 #endif
